requirements.txt
streamlit==1.29.0
tensorflow==2.12.0
numpy==1.23.5
pandas==2.0.2
scikit-learn==1.2.2
matplotlib==3.7.1
opencv-python==4.7.0.72
Pillow==9.5.0